Fix: sprite-timeline can not transplied by typescript-plugin
