+++ packages/eva.js/lib/timeline/utils.ts
@@ -1,0 +1,10 @@
+export function createNowTime() {
+  let nowtime = null;
+  if (Date.now) {
+    nowtime = Date.now;
+  } else {
+    nowtime = () => new Date().getTime();
+  }
+
+  return nowtime;
+}
